Move \Slim\Collection beneath \Slim\Http namespace
Solves #1142
