+++ tests/MiddlewareAwareTest.php
@@ -59,8 +59,8 @@ class MiddlewareTest extends \PHPUnit_Framework_TestCase
         // Request
         $uri = \Slim\Http\Uri::createFromString('https://example.com:443/foo/bar?abc=123');
         $headers = new \Slim\Http\Headers();
-        $cookies = new \Slim\Collection();
-        $serverParams = new \Slim\Collection();
+        $cookies = new \Slim\Http\Collection();
+        $serverParams = new \Slim\Http\Collection();
         $body = new \Slim\Http\Body(fopen('php://temp', 'r+'));
         $request = new \Slim\Http\Request('GET', $uri, $headers, $cookies, $serverParams, $body);
 
@@ -91,8 +91,8 @@ class MiddlewareTest extends \PHPUnit_Framework_TestCase
         // Request
         $uri = \Slim\Http\Uri::createFromString('https://example.com:443/foo/bar?abc=123');
         $headers = new \Slim\Http\Headers();
-        $cookies = new \Slim\Collection();
-        $serverParams = new \Slim\Collection();
+        $cookies = new \Slim\Http\Collection();
+        $serverParams = new \Slim\Http\Collection();
         $body = new \Slim\Http\Body(fopen('php://temp', 'r+'));
         $request = new \Slim\Http\Request('GET', $uri, $headers, $cookies, $serverParams, $body);
